Удалил лишние ветки с меню. Последний класс перенес в основную ветку.
